requirements.txt
numpy
sklearn
tqdm
geopandas
requests
scikit-learn
fastapi
uvicorn[standard]
geojson-pydantic
h3
azure-storage-file-datalake
